MXMD reading works
Additional changes:

* Added base_progress_reporter::forward() to forward an existing bound progress reporter function. We don't have to create as many function objects this way

* FS path objects are now const refs. If we need to make changes we will make a temporary copy when we need it.

+++ src/core/ivstream.h
@@ -26,7 +26,7 @@ namespace core {
 
 			pos_type seekoff(off_type off, std::ios_base::seekdir dir, std::ios_base::openmode which = std::ios_base::in) override {
 				if (dir == std::ios_base::cur)
-					gbump(off);
+					gbump((int32)off);
 				else if (dir == std::ios_base::end)
 					setg(eback(), egptr() + off, egptr());
 				else if (dir == std::ios_base::beg)
@@ -57,7 +57,7 @@ namespace core {
 			int_type pbackfail(int_type c) {
 				// if they are trying to push back a character that they didn't read last
 				// that is an error
-				const unsigned long prev = read_pos-1;
+				const unsigned long prev = (unsigned long)read_pos-1;
 				if (c != EOF && prev < buffer.size() && 
 					c != static_cast<unsigned char>(buffer[prev])) {
 					return EOF;
